将玩家npc怪物归类于object 修改定时器为线程

+++ works/nettyServer/src/main/java/com/function/monster/excel/MonsterExcel.java
@@ -34,7 +34,7 @@ public class MonsterExcel {
     /**
      * 复活时间
      */
-    private int reviveTime;
+    private Long reviveTime;
     /**
      * 掉落
      */
